[mlir] enable yolo_v4 bf16
Symptom: NA
Root cause: NA
Solution: NA
Note:
1. rewrite mish to x * tanh() for high accuracy

+++ lib/Dialect/TPU/Backend/TpuTlCodegen_Simple.cpp
@@ -45,8 +45,8 @@
 
 using namespace mlir;
 
-extern int BF16_TABLE_START;
-extern int BF16_TABLE_END;
+extern float BF16_TABLE_START;
+extern float BF16_TABLE_END;
 
 #include "backend/backend_tg_api.h"
 #include "backend/backend_tl_api.h"
